Implement issue #29 - track number can be a floating point number now.
The parser is a bit hacky but would work.

A remaining issue to note is that track range specification won't work
well with it, because any track number can be assigned later and we cannot
cover them in prior i.e.

```
1-100 r1

1	c2
1.1	e2
```

This `1.1` cannot be covered because we have no idea whether a number
(among infinite number of double values) between 0 and 1 is covered by
this or not, or allocate it "for any possible number that can appear later".

Thus, track range specification works only as an "incremented values
in integer" i.e. 1-10 means 1, 2, 3, 4, 5, 6, 7, 8, 9, and 10.

+++ mugenelib/src/mml_parser.cs
@@ -756,7 +756,7 @@ void case_47()
 {
 		var t = (MmlToken) yyVals[-1+yyTop];
 		var d = (int) yyVals[0+yyTop];
-		yyVal = new MmlConstantExpr (t.Location, MmlDataType.Length, new MmlLength ((int) t.Value) { Dots = d });
+		yyVal = new MmlConstantExpr (t.Location, MmlDataType.Length, new MmlLength ((int) (double) t.Value) { Dots = d });
 	}
 
 void case_48()
